Add test for preload.

+++ regress/cat.c
@@ -23,7 +23,7 @@ int main(int argc, char *argv[]) {
 	    continue;
 	}
 	while ((ret=fread(buf, 1, sizeof(buf), in)) > 0) {
-	    printf("%.*s", ret, buf);
+	    printf("%.*s", (int)ret, buf);
 	}
 	if (in != stdin) {
 	    fclose(in);
